bitcoin: move sighash tests to tests/
Move legacy_sighash and bip_341_sighash_tests to
tests/ because they include_str data from tests/, which
we exclude when publishing.

Cache assertions in bip_341_sighash_tests call private
common_cache and taproot_cache, so they stay a unit test
with the first keyPathSpending vector of
bip341_tests.json inlined.
